Split BMI088 part of sensors_bmi088_bmp388 into backends
Instead of having one complete driver for each BMI088 interface we split
the BMI088 part into a I2C and a SPI backend. This way we have one
driver implementing functionality and handling LOG and PARAM and the
backends take care of interface details.

New structure:

 sensors.c => tells us which bmi088 init function (SPI/I2C) to use in ...
   sensors_bmi088_bmp388.c => which defers interface init to ...
     => sensors_bmi088_[i2c|spi].c => which sets up device information.

Two files are removed:
  sensors_bmi088_spi_bmp388.c
  sensors_bmi088_spi_bmp388.h

Three files are added:
  sensors_bmi088_i2c.c     - unique i2c device init
  sensors_bmi088_spi.c     - unique spi functions and evice init
  sensors_bmi088_common.h  - shared function definitions
