+++ tools/make/platforms/cf2.mk
@@ -6,14 +6,12 @@ PLATFORM_NAME_cf2 = CF2 platform
 CPU=stm32f4
 
 ######### Sensors configuration ##########
-CFLAGS += -DSENSOR_INCLUDED_BMI088_BMP388
-PROJ_OBJ += sensors_bmi088_bmp388.o
-
 CFLAGS += -DSENSOR_INCLUDED_MPU9250_LPS25H
 PROJ_OBJ += sensors_mpu9250_lps25h.o
 
+CFLAGS += -DSENSOR_INCLUDED_BMI088_BMP388
 CFLAGS += -DSENSOR_INCLUDED_BMI088_SPI_BMP388
-PROJ_OBJ += sensors_bmi088_spi_bmp388.o
+PROJ_OBJ += sensors_bmi088_bmp388.o sensors_bmi088_i2c.o sensors_bmi088_spi.o
 
 ######### Stabilizer configuration ##########
 ESTIMATOR          ?= any
